created useIsGovMember hooks and using it
